separating Client and Service classes
